requirements.txt
streamlit>=1.32.0
anthropic>=0.25.0
pandas>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
